debugged a little bit

+++ factorial.cpp
@@ -1,5 +1,5 @@
-int factorial(int num){
-	int fact = num;
+long long factorial(int num){
+	long long fact = num;
 	for(int i=num-1; i>1; i--){
 		fact *= i;
 	}
