+++ power.cpp
@@ -1,6 +1,6 @@
-int power(double base, int exponent){
+long double power(double base, int exponent){
 	int n = exponent-1;
-	int result = base;
+	long double result = base;
 	for(int i=0; i<n; i++){
 		result *= base;
 	}
